+++ backend/src/main/java/com/sims/admin/controller/AdminController.java
@@ -2,15 +2,16 @@ package com.sims.admin.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@RestController
+@Controller
 public class AdminController {
 
-    @GetMapping("/admin")
-    public ResponseEntity<String> admin() {
+    @GetMapping("/main")
+    public String admin() {
 
-        return new ResponseEntity<>("Test2", HttpStatus.OK);
+        return "sidebar/sidebar";
     }
 }
